Refactor portfolio into multiple Astro pages and split React modules
- Add static routes for /, /web, /games, /blog, /about, /now, /contact, /resume
- Extract global styles to src/styles/global.css and shared BaseLayout.astro
- Replace monolithic PortfolioApp with portfolio/ (shell, sections, modals, SVGs)
- Use react-router for client navigation; keep curtain transition and theme
- Client-side Sanity fetch preserved so static build works without projectId at build time
