+++ package.json
@@ -15,6 +15,8 @@
   "dependencies": {
     "@astrojs/react": "^5.0.1",
     "@astrojs/tailwind": "^6.0.2",
+    "@sanity/client": "^6.0.0",
+    "@sanity/vision": "^3.0.0",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "^0.5.10",
     "@types/react": "^19.2.14",
@@ -22,9 +24,8 @@
     "astro": "^5.16.5",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
-    "tailwindcss": "^3.4.19",
-    "@sanity/client": "^6.0.0",
+    "react-router-dom": "^7.14.1",
     "sanity": "^3.0.0",
-    "@sanity/vision": "^3.0.0"
+    "tailwindcss": "^3.4.19"
   }
 }
