+++ src/pages/contact.astro
@@ -3,6 +3,6 @@ import BaseLayout from "../layouts/BaseLayout.astro";
 import PortfolioShell from "../components/portfolio/PortfolioShell";
 ---
 
-<BaseLayout>
+<BaseLayout title="Contact — Jon Colon — Designer, Developer, Game Dev">
   <PortfolioShell client:only="react" />
 </BaseLayout>
